Move Varnish stuff out into its own app
