+++ esp/esp/django_settings.py
@@ -222,6 +222,7 @@ INSTALLED_APPS = (
     'esp.seltests',
     'esp.dataviews',
     'esp.themes',
+    'esp.varnish',
     'django_extensions',
     'reversion',
     'south',
