+++ esp/esp/varnish/__init__.py
@@ -1,0 +1,1 @@
+from esp.varnish.varnish import purge_page
