+++ esp/esp/varnish/urls.py
@@ -1,5 +1,5 @@
 from django.conf.urls.defaults import *
 
 urlpatterns = patterns('',
-                        (r'^view_all/?$', 'esp.cache.views.view_all'),
+                        (r'^purge_page$', 'esp.varnish.views.varnish_purge'),
                         )
